+++ spec/lib/dawn/codesake_knowledgebase_spec.rb
@@ -988,4 +988,9 @@ describe "The Codesake Dawn knowledge base" do
     sc.class.should == Dawn::Kb::CVE_2015_2963
   end
   it "CVE-2015-1840 can't be tested. Trust me. It works"
+  it "must have test for CVE-2015-3448" do
+    sc = kb.find("CVE-2015-3448")
+    sc.should_not   be_nil
+    sc.class.should == Dawn::Kb::CVE_2015_3448
+  end
 end